+++ lib/sequential/sequential.h
@@ -1,0 +1,6 @@
+#ifndef HW_C_PLUS_PLUS_SEQUENTIAL_H
+#define HW_C_PLUS_PLUS_SEQUENTIAL_H
+
+int sequential(const char *filename, long size, long result[2]);
+
+#endif //HW_C_PLUS_PLUS_SEQUENTIAL_H
