+++ tests/CMakeLists.txt
@@ -1,9 +1,12 @@
 cmake_minimum_required(VERSION 3.16)
 
-set(PROJECT_SOURCE_FILES)
+#set(PROJECT_SOURCE_FILES)
 
 add_executable(tests test.cpp ${PROJECT_SOURCE_FILES})
 
-include_directories(../project_consistently/include)
+include_directories(../include)
+include_directories(../lib/parallel)
+include_directories(../lib/sequential)
 
 target_link_libraries(tests ${GTEST_LIBRARIES} pthread)
+target_link_libraries(tests sequential_lib parallel_lib)
